Breaking up llvm-ir2vec lib implementation to clean up MIR deps from ir2vec python bindings
