+++ llvm/tools/llvm-ir2vec/Bindings/PyIR2Vec.cpp
@@ -6,7 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "lib/Utils.h"
+#include "lib/IRUtils.h"
 #include "llvm/IR/LLVMContext.h"
 #include "llvm/IR/Module.h"
 #include "llvm/IRReader/IRReader.h"
