+++ llvm/tools/llvm-ir2vec/lib/CMakeLists.txt
@@ -1,6 +1,7 @@
-add_llvm_library(LLVMEmbUtils STATIC
-  Utils.cpp
+add_llvm_library(LLVMIREmbUtils STATIC
+  IRUtils.cpp
 
+  PARTIAL_SOURCES_INTENDED
   BUILDTREE_ONLY
 
   DEPENDS
@@ -8,8 +9,20 @@ add_llvm_library(LLVMEmbUtils STATIC
 
   LINK_COMPONENTS
   Analysis
-  CodeGen
   Core
   Support
+)
+
+add_llvm_library(LLVMMIREmbUtils STATIC
+  MIRUtils.cpp
+
+  PARTIAL_SOURCES_INTENDED
+  BUILDTREE_ONLY
+
+  DEPENDS
+  intrinsics_gen
+
+  LINK_COMPONENTS
+  CodeGen
   Target
 )
